+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 Python class for reading and writing
 [NonLinLoc](http://alomax.free.fr/nlloc) grid files.
 
-(c) 2015-2023 Claudio Satriano, Natalia Poiata, Robert Pickle
+(c) 2015-2024 Claudio Satriano, Natalia Poiata, Robert Pickle
 
 ## unreleased